requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
openai==1.68.*
pydantic-settings==2.7.*
python-dotenv==1.0.*
tavily-python>=0.5.*
